% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcul_distribution_age.R
\name{calcul_distribution_age}
\alias{calcul_distribution_age}
\title{Calculer la distribution des âges des élus}
\usage{
calcul_distribution_age(df)
}
\arguments{
\item{df}{Un `data.frame` contenant les données des élus, avec la colonne `Date.de.naissance` utilisée
pour calculer l'âge de chaque élu.}
}
\value{
Une liste contenant les quantiles des âges des élus :
- `Q0` : 0% (minimum)
- `Q25` : 25% (premier quartile)
- `Q50` : 50% (médiane)
- `Q75` : 75% (troisième quartile)
- `Q100` : 100% (maximum)
}
\description{
Cette fonction permet de calculer la distribution des âges des élus à partir de leur date de naissance.
Elle retourne les quantiles (0%, 25%, 50%, 75%, 100%) des âges calculés en années.
}
