% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compter_nombre_d_adjoints.R
\name{compter_nombre_d_adjoints}
\alias{compter_nombre_d_adjoints}
\title{Compter le nombre d'adjoints dans un DataFrame}
\usage{
compter_nombre_d_adjoints(df)
}
\arguments{
\item{df}{Un `data.frame` contenant les données des élus. Il doit inclure la colonne
`Libellé.de.la.fonction` qui sera utilisée pour rechercher le terme "adjoint".}
}
\value{
Un entier représentant le nombre d'adjoints dans le `data.frame`.
}
\description{
Cette fonction permet de compter le nombre de lignes dans un `data.frame` où la fonction
de l'élu contient le mot "adjoint". Elle utilise la fonction `validate_schema` pour vérifier
que le `data.frame` respecte la structure minimale.
}
