% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compter_nombre_d_elus.R
\name{compter_nombre_d_elus}
\alias{compter_nombre_d_elus}
\title{Compter le nombre d'élus dans un DataFrame}
\usage{
compter_nombre_d_elus(df)
}
\arguments{
\item{df}{Un `data.frame` contenant les données des élus, avec les colonnes `Nom.de.l.élu`,
`Prénom.de.l.élu` et `Date.de.naissance` utilisées pour déterminer l'unicité de chaque élu.}
}
\value{
Un entier représentant le nombre d'élus uniques dans le `data.frame`.
}
\description{
Cette fonction permet de compter le nombre d'élus uniques dans un `data.frame` en fonction de leur
nom, prénom et date de naissance. Elle élimine les doublons pour ne compter qu'une seule fois
chaque élu, même s'il apparaît plusieurs fois dans le `data.frame`.
}
