% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/creer_departement.R
\name{creer_departement}
\alias{creer_departement}
\title{Créer un objet de classe "departement"}
\usage{
creer_departement(df)
}
\arguments{
\item{df}{Un `data.frame` contenant les données des départements, avec la colonne `Code.du.département` utilisée
pour vérifier si plusieurs départements sont présents.}
}
\value{
Un `data.frame` auquel la classe "departement" est ajoutée s'il contient un seul département.
}
\description{
Cette fonction vérifie si le `data.frame` contient les données d'un seul département. Si plusieurs départements sont détectés,
elle arrête l'exécution avec un message d'erreur. Sinon, elle ajoute la classe "departement" au `data.frame`.
}
