% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/firstlibdjayan-package.R
\docType{package}
\name{firstlibdjayan}
\alias{firstlibdjayan-package}
\alias{firstlibdjayan}
\title{firstlibdjayan: Analyse et Visualisation des Données Communales et Départementales}
\description{
Ce package fournit des fonctions pour manipuler les données relatives aux élus des communes
et départements français. Il inclut des outils pour calculer des statistiques, générer des graphiques
et structurer les données en classes `commune` et `departement`.
}
\details{
Ce package permet l'analyse et la visualisation des données relatives aux élus des communes
et départements français. Il inclut des outils pour calculer des statistiques, générer des graphiques
et structurer les données en classes `commune` et `departement`.
}
\seealso{
\url{https://github.com/Djayan-D/firstlibdjayan/} pour accéder au code source du package.
\url{https://djayan-d.github.io/firstlibdjayan/} pour consulter la documentation en ligne.
\url{https://sites.google.com/view/master-ecap/accueil} pour plus d'informations sur le Master ECAP.
}
\author{
Daëron Djayan [aut, cre] (<daeron.djayan@gmail.com>)
}
