% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_commune.R
\name{plot.commune}
\alias{plot.commune}
\title{Graphique des codes professionnels des élus d'une commune}
\usage{
\method{plot}{commune}(df)
}
\arguments{
\item{df}{Un objet de classe \code{commune} (un \code{data.frame} contenant les informations des élus de la commune).}
}
\value{
Un objet de type \code{ggplot} représentant un graphique en barres horizontales, où l'axe des ordonnées correspond aux codes professionnels
  et l'axe des abscisses au nombre d'élus. Le graphique inclut également des étiquettes des valeurs et un titre personnalisé.
}
\description{
Cette fonction génère un graphique en barres horizontales représentant le nombre d'élus par code professionnel dans une commune.
Le graphique affiche les codes professionnels ayant au moins un élu, triés par nombre décroissant d'élus. Un titre et des étiquettes
sont ajoutés pour la lisibilité.
}
\examples{
# Créer un exemple de jeu de données pour Nantes (code 44109)
df_commune <- data.frame(
  Code.de.la.commune = rep(44109, 10),
  Libellé.de.la.commune = rep("Nantes", 10),
  Libellé.du.département = rep("Loire-Atlantique", 10),
  Code.de.la.catégorie.socio.professionnelle = c(1, 2, 2, 3, 1, 3, 4, 1, 2, 3),
  stringsAsFactors = FALSE
)
df_commune <- creer_commune(df_commune)  # Ajouter la classe "commune"

# Appeler la fonction pour générer le graphique
plot(df_commune)

}
\seealso{
\itemize{
  \item \href{https://djayan-d.github.io/firstlibdjayan/}{Documentation sur la création de DataFrame}
  \item \href{https://github.com/Djayan-D/firstlibdjayan/}{Code source sur GitHub}
  \item \href{https://djayan-d.github.io/firstlibdjayan/}{Documentation en ligne}
  }
}
