% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_departement.R
\name{plot.departement}
\alias{plot.departement}
\title{Graphique des codes professionnels des élus d'un département}
\usage{
\method{plot}{departement}(df)
}
\arguments{
\item{df}{Un objet de classe \code{departement} (un \code{dataframe} contenant les informations des élus du département).}
}
\value{
Un objet de type \code{ggplot} représentant un graphique en barres horizontales, où l'axe des ordonnées correspond aux codes professionnels
  et l'axe des abscisses au nombre d'élus. Le graphique inclut également des étiquettes des valeurs et un titre personnalisé.
}
\description{
Cette fonction génère un graphique en barres horizontales représentant les 10 codes professionnels les plus représentés parmi les élus
d'un département. Le graphique affiche les codes professionnels ayant au moins un élu, triés par nombre décroissant d'élus.
Le titre du graphique et l'axe des abscisses sont personnalisés en fonction du nom du département et du nombre de communes.
}
