% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_code_professions.R
\name{plot_code_professions}
\alias{plot_code_professions}
\title{Générer un graphique du nombre d'élus par code professionnel}
\usage{
plot_code_professions(df)
}
\arguments{
\item{df}{Un `data.frame` contenant les données des élus, avec la colonne `Code.de.la.catégorie.socio.professionnelle` utilisée
pour déterminer la répartition des élus par code professionnel.}
}
\value{
Un objet `ggplot` représentant un graphique en barres horizontal du nombre d'élus par code professionnel.
}
\description{
Cette fonction permet de générer un graphique en barres horizontal représentant le nombre d'élus
par code professionnel. Elle compte le nombre d'élus par code professionnel, filtre ceux avec zéro élu
et affiche les résultats sous forme de graphique.
}
