% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trouver_communes_extremes_age.R
\name{trouver_communes_extremes_age}
\alias{trouver_communes_extremes_age}
\title{Trouver les communes avec les âges extrêmes}
\usage{
trouver_communes_extremes_age(df)
}
\arguments{
\item{df}{Un `data.frame` contenant les informations des élus, avec la colonne `Date.de.naissance` utilisée pour calculer
la moyenne d'âge par commune.}
}
\value{
Une liste contenant deux sous-listes :
- `commune_plus_jeune` : La commune avec la moyenne d'âge la plus faible, incluant sa distribution d'âges,
- `commune_plus_agee` : La commune avec la moyenne d'âge la plus élevée, incluant sa distribution d'âges.
}
\description{
Cette fonction permet d'identifier les communes ayant la moyenne d'âge la plus faible et la plus élevée parmi les élus.
Elle calcule la moyenne d'âge pour chaque commune, puis retourne une liste contenant les informations suivantes pour
les communes avec les âges les plus jeunes et les plus âgés :
- Le nom de la commune,
- La distribution des âges des élus.
}
