% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trouver_l_elu_le_plus_age.R
\name{trouver_l_elu_le_plus_age}
\alias{trouver_l_elu_le_plus_age}
\title{Trouver l'élu le plus âgé}
\usage{
trouver_l_elu_le_plus_age(df)
}
\arguments{
\item{df}{Un `data.frame` contenant les données des élus, avec les colonnes `Nom.de.l.élu`,
`Prénom.de.l.élu` et `Date.de.naissance` nécessaires pour déterminer l'élu le plus âgé.}
}
\value{
Une liste contenant les informations suivantes sur l'élu le plus âgé :
- `nom` : Le nom de l'élu
- `prénom` : Le prénom de l'élu
- `age` : L'âge de l'élu en années
}
\description{
Cette fonction permet de trouver l'élu le plus âgé dans un `data.frame`. Elle trie les élus par
date de naissance et sélectionne l'élu ayant la date de naissance la plus ancienne. L'âge de l'élu
est ensuite calculé en années.
}
