% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trouver_l_elu_le_plus_jeune.R
\name{trouver_l_elu_le_plus_jeune}
\alias{trouver_l_elu_le_plus_jeune}
\title{Trouver l'élu le plus jeune}
\usage{
trouver_l_elu_le_plus_jeune(df)
}
\arguments{
\item{df}{Un `data.frame` contenant les informations des élus, avec la colonne `Date.de.naissance` utilisée pour
identifier l'élu le plus jeune.}
}
\value{
Une liste contenant :
- `nom` : Le nom de l'élu le plus jeune,
- `prénom` : Le prénom de l'élu le plus jeune,
- `age` : L'âge de l'élu le plus jeune en années.
}
\description{
Cette fonction permet d'extraire les informations de l'élu le plus jeune d'un `data.frame`. L'élu le plus jeune est
identifié en fonction de sa date de naissance la plus récente. La fonction retourne un résumé des informations de cet élu,
incluant son nom, prénom et son âge calculé en années.
}
