% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/creer_commune.R
\name{creer_commune}
\alias{creer_commune}
\title{Créer un objet de classe \code{commune}}
\usage{
creer_commune(df)
}
\arguments{
\item{df}{Un \code{data.frame} contenant les données des communes, avec la colonne \code{Code.de.la.commune} utilisée
pour vérifier si plusieurs communes sont présentes.}
}
\value{
Un \code{data.frame} auquel la classe \code{commune} est ajoutée s'il contient une seule commune.
}
\description{
Cette fonction vérifie si le \code{data.frame} contient les données d'une seule commune. Si plusieurs communes sont détectées,
elle arrête l'exécution avec un message d'erreur. Sinon, elle ajoute la classe \code{commune} au \code{data.frame}.
}
