% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/creer_commune.R
\docType{data}
\name{creer_commune}
\alias{creer_commune}
\title{Créer un objet de classe \code{commune}}
\format{
An object of class \code{function} of length 1.
}
\usage{
creer_commune
}
\arguments{
\item{df}{Un \code{data.frame} contenant les données des communes, avec la colonne \code{Code.de.la.commune} utilisée
pour vérifier si plusieurs communes sont présentes.}
}
\value{
Un \code{data.frame} auquel la classe \code{commune} est ajoutée s'il contient une seule commune.
}
\description{
Cette fonction vérifie si le \code{data.frame} contient les données d'une seule commune. Si plusieurs communes sont détectées,
elle arrête l'exécution avec un message d'erreur. Sinon, elle ajoute la classe \code{commune} au \code{data.frame}.
}
\examples{
Exemple de données pour une commune

df_commune <- data.frame(Code.de.la.commune = c(44109,44109))
df_commune <- creer_commune(df_commune)
class(df_commune)  # Doit afficher "commune" comme classe ajoutée

}
\seealso{
\itemize{
  \item \href{https://djayan-d.github.io/firstlibdjayan/}{Documentation sur la création de DataFrame}
  \item \href{https://github.com/Djayan-D/firstlibdjayan/}{Code source sur GitHub}
  \item \href{https://djayan-d.github.io/firstlibdjayan/}{Documentation en ligne}
  }
}
\keyword{datasets}
