% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/creer_departement.R
\name{creer_departement}
\alias{creer_departement}
\title{Créer un objet de classe \code{departement}}
\usage{
creer_departement(df)
}
\arguments{
\item{df}{Un \code{data.frame} contenant les données des départements, avec la colonne \code{Code.du.département} utilisée
pour vérifier si plusieurs départements sont présents.}
}
\value{
Un \code{data.frame} auquel la classe \code{departement} est ajoutée s'il contient un seul département.
}
\description{
Cette fonction vérifie si le \code{data.frame} contient les données d'un seul département. Si plusieurs départements sont détectés,
elle arrête l'exécution avec un message d'erreur. Sinon, elle ajoute la classe \code{departement} au \code{data.frame}.
}
