% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/creer_departement.R
\docType{data}
\name{creer_departement}
\alias{creer_departement}
\title{Créer un objet de classe \code{departement}}
\format{
An object of class \code{function} of length 1.
}
\usage{
creer_departement
}
\arguments{
\item{df}{Un \code{data.frame} contenant les données des départements, avec la colonne \code{Code.du.département} utilisée
pour vérifier si plusieurs départements sont présents.}
}
\value{
Un \code{data.frame} auquel la classe \code{departement} est ajoutée s'il contient un seul département.
}
\description{
Cette fonction vérifie si le \code{data.frame} contient les données d'un seul département. Si plusieurs départements sont détectés,
elle arrête l'exécution avec un message d'erreur. Sinon, elle ajoute la classe \code{departement} au \code{data.frame}.
}
\examples{
#Exemple de données pour un département

df_departement <- data.frame(Code.du.département = c(44,44))
df_departement <- creer_departement(df_departement)
class(df_departement)  # Doit afficher "departement" comme classe ajoutée

}
\seealso{
\itemize{
  \item \href{https://djayan-d.github.io/firstlibdjayan/}{Documentation sur la création de DataFrame}
  \item \href{https://github.com/Djayan-D/firstlibdjayan/}{Code source sur GitHub}
  \item \href{https://djayan-d.github.io/firstlibdjayan/}{Documentation en ligne}
  }
}
\keyword{datasets}
