% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/firstlibdjayan-package.R
\docType{package}
\name{firstlibdjayan}
\alias{firstlibdjayan-package}
\alias{firstlibdjayan}
\title{firstlibdjayan: Analyse et Visualisation des Données Communales et Départementales}
\description{
Ce package fournit des fonctions pour manipuler les données relatives aux élus des communes
et départements français. Il inclut des outils pour calculer des statistiques, générer des graphiques
et structurer les données en classes \code{commune} et \code{departement}.
}
\details{
Ce package permet l'analyse et la visualisation des données relatives aux élus des communes
et départements français. Il inclut des outils pour calculer des statistiques, générer des graphiques
et structurer les données en classes \code{commune} et \code{departement}.
}
\seealso{
\itemize{
  \item \href{https://github.com/Djayan-D/firstlibdjayan/}{Code source sur GitHub}
  \item \href{https://djayan-d.github.io/firstlibdjayan/}{Documentation en ligne}
  \item \href{https://sites.google.com/view/master-ecap/accueil}{Master ECAP}
}
}
\author{
Daëron Djayan [aut, cre] \email{daeron.djayan@gmail.com}
}
