% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generer_rapport.R
\name{generer_rapport}
\alias{generer_rapport}
\title{Générer un rapport à partir d'un fichier Quarto}
\usage{
<<<<<<< HEAD
generer_rapport(commune, departement, output, df = NULL)
=======
generer_rapport(commune, departement, output)
>>>>>>> gh-pages
}
\arguments{
\item{commune}{Le code (ex : INSEE) associé à la commune pour laquelle le rapport sera généré.}

\item{departement}{Le code (ex : INSEE) associé au département pour lequel le rapport sera généré.}

<<<<<<< HEAD
\item{output}{Un caractère spécifiant le chemin du fichier de sortie (par exemple, "rapport_nantes.html").}
}
\value{
Aucun. La fonction génère un fichier de rapport au format spécifié (par défaut html).
=======
\item{output}{Un caractère spécifiant le chemin du fichier de sortie (par exemple, "rapport_nantes.pdf").}
}
\value{
Aucun. La fonction génère un fichier de rapport au format spécifié (par défaut PDF).
>>>>>>> gh-pages
}
\description{
Cette fonction génère un rapport au format PDF à partir d'un fichier Quarto (.qmd) en utilisant les paramètres de la commune et du département fournis.
}
\examples{
<<<<<<< HEAD
generer_rapport(44109, 44, "rapport_Nantes_et_Loire_Atlantique.html")
=======
generer_rapport(44109, 44, "rapport_Nantes_et_Loire_Atlantique.pdf")
>>>>>>> gh-pages
}
