% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generer_rapport.R
\name{generer_rapport}
\alias{generer_rapport}
\title{Générer un rapport à partir d'un fichier Quarto}
\usage{
generer_rapport(commune, departement, output)
}
\arguments{
\item{commune}{Un caractère représentant le nom de la commune pour laquelle le rapport sera généré.}

\item{departement}{Un caractère représentant le nom du département pour lequel le rapport sera généré.}

\item{output}{Un caractère spécifiant le chemin du fichier de sortie (par exemple, "rapport_nantes.pdf").}
}
\value{
Aucun. La fonction génère un fichier de rapport au format spécifié (par défaut PDF).
}
\description{
Cette fonction génère un rapport au format PDF à partir d'un fichier Quarto (.qmd) en utilisant les paramètres de la commune et du département fournis.
}
\examples{
generer_rapport("Nantes", "Loire-Atlantique", "rapport_nantes.pdf")
}
